fix up examples
