+++ example/cache/in-memory/main.go
@@ -20,58 +20,10 @@ func main() {
 
 	for i := 0; i < 2; i++ {
 		fmt.Println([]string{"Actual requests", "From cache"}[i])
-		if v, err := ipinfo.GetIP(ip); err != nil {
+		if v, err := ipinfo.GetIpInfo(ip); err != nil {
 			log.Println(err)
 		} else {
-			fmt.Println("IP:", v)
-		}
-
-		if v, err := ipinfo.GetHostname(ip); err != nil {
-			log.Println(err)
-		} else {
-			fmt.Println("Hostname:", v)
-		}
-
-		if v, err := ipinfo.GetOrganization(ip); err != nil {
-			log.Println(err)
-		} else {
-			fmt.Println("Organization:", v)
-		}
-
-		if v, err := ipinfo.GetCity(ip); err != nil {
-			log.Println(err)
-		} else {
-			fmt.Println("City:", v)
-		}
-
-		if v, err := ipinfo.GetRegion(ip); err != nil {
-			log.Println(err)
-		} else {
-			fmt.Println("Region:", v)
-		}
-
-		if v, err := ipinfo.GetCountry(ip); err != nil {
-			log.Println(err)
-		} else {
-			fmt.Println("Country:", v)
-		}
-
-		if v, err := ipinfo.GetLocation(ip); err != nil {
-			log.Println(err)
-		} else {
-			fmt.Println("Location:", v)
-		}
-
-		if v, err := ipinfo.GetPhone(ip); err != nil {
-			log.Println(err)
-		} else {
-			fmt.Println("Phone:", v)
-		}
-
-		if v, err := ipinfo.GetPostal(ip); err != nil {
-			log.Println(err)
-		} else {
-			fmt.Println("Postal:", v)
+			fmt.Printf("IP: %v\n", v)
 		}
 	}
 }
