+++ example/my-ip-info/main.go
@@ -8,14 +8,9 @@ import (
 )
 
 func main() {
-	info, err := ipinfo.GetInfo(nil)
+	core, err := ipinfo.GetIpInfo(nil)
 	if err != nil {
 		log.Fatal(err)
 	}
-	printInfo(info)
-}
-
-func printInfo(info *ipinfo.Info) {
-	fmt.Printf("IP: %v\nHostname: %s\nOrganization: %s\nCity: %s\nRegion: %s\nCountry: %s\nLocation: %s\nPhone: %s\nPostal: %s\n",
-		info.IP, info.Hostname, info.Organization, info.City, info.Region, info.Country, info.Location, info.Phone, info.Postal)
+	fmt.Printf("%v\n", core)
 }
